+++ src/FluentDbTools/Abstractions/FluentDbTools.Common.Abstractions/ValidateDatabaseAdminValuesExtensions.cs
@@ -1,13 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using FluentDbTools.Contracts;
 #pragma warning disable CS1591
 
+[assembly: InternalsVisibleTo("FluentDbTools.Extensions.DbProvider")]
 namespace FluentDbTools.Common.Abstractions
 {
     public static class ValidateDatabaseAdminValuesExtensions
     {
+        internal const string OneOfTheRequiredConfigurationParameters = "One of the required configuration parameters";
+
+        internal const string AllRequiredConfigurationParameters = "All required configuration parameters";
+
+        internal const string RequiredConfigurationParameter = "Required configuration parameter";
+
         /// <summary>
         /// Convert from "section:section2:value" to "'Section.Section2.Value'";
         /// </summary>
@@ -58,8 +66,33 @@ namespace FluentDbTools.Common.Abstractions
 
         public static bool IsInvalidDatabaseAdminException(this Exception exception)
         {
-            return exception.Source == InvalidAdminType.AdminUser.ToString("G") ||
+            if (exception == null)
+            {
+                return false;
+            }
+
+            var isInvalidException =
+                   exception.Source == InvalidAdminType.AdminUser.ToString("G") ||
                    exception.Source == InvalidAdminType.AdminPassword.ToString("G");
+
+            if (isInvalidException)
+            {
+                return true;
+            }
+
+            var messagesContainsStrings = new[] { OneOfTheRequiredConfigurationParameters, AllRequiredConfigurationParameters, RequiredConfigurationParameter };
+
+            switch (exception)
+            {
+                case ArgumentNullException argumentNullException:
+                    return argumentNullException.ParamName != null && messagesContainsStrings.Any(x => argumentNullException.Message.ContainsIgnoreCase(x));
+                case ArgumentException argumentException:
+                    return messagesContainsStrings.Any(x => argumentException.Message.ContainsIgnoreCase(x));
+                case AggregateException aggregateException:
+                    return aggregateException.InnerExceptions.Any(innerException => messagesContainsStrings.Any(x => IsInvalidDatabaseAdminException(innerException)));
+                default:
+                    return false;
+            }
         }
     }
 }